Add `max_buffers` options to disable buffer tabs
This will disable the buffer tabs if the amount of buffers exceed this number, this is helpful if you have a lot of buffers as they don't all fit on the screen.

I don't really want any more than about 6 buffer tabs at a time.

This value defaults to 0 which is current behaviour.