Add `surround_active_buffer` option to limit number of buffer tabs
This option only shows a buffer tab for the active buffer and `surround_active_buffer` number of buffer tabs either side.

This is helpful when you have too many buffers open to fit tabs on the screen but means you can still see what are likely to be the most important ones (those that are close by).

This defaults to 0 which is current behaviour.

If there are not enough buffers to have a tab for the current buffer and `surround_active_buffer` number of tabs either side we fall back to default behaviour.