Fix rendering without icons
This fixes the issue where the text doesn't fit in the buffer tab window when icons are set to false, this was happening because we weren't respecting utf8 length of strings.

This commit switches to using `vim.fn.strdisplaywidth` to get the actual character length of the string.

This also remedies some un-needed whitespace that is inserted when icons are disabled.